requirements.txt
anthropic>=0.20.0
pydantic>=2.0
rich>=13.0
pytest>=7.0
pytest-asyncio>=0.21
python-dotenv>=1.0
